Try to fail more gracefully when saving files
If the minor mode is active, we will always try to make a commit when
the buffer is saved. Depending on the setting of `pcmm-commit-all` and
the state of the buffer (whether it is associated with a file, and
whether it is already tracked by git), you should receive different
error messages.