Prevent buffer save prompts from magit
This patch prevents interrupting the user with "save such-and-such
buffer?" prompts by no longer calling `magit-stage-modified` and
instead calling `vc-git-command` directly to stage unstaged files when
`pcmm-commit-all` is non-nil.

In the case where `pcmm-commit-all` is nil, we still commit all
tracked changes by adding the `a` flag to the commit command.